Added bitmap.cpp: Modified from last.
Fergus' code from before, but changed the return type to the array_data struct.

Also added a boolean mask, which is used in the FDM function to 'protect' boundary values. Much smaller in memory than a duplicate array since it only holds 0 or 1, and should be quicker than testing and rewriting each boundary element.

ar -> values[p][q] is similar to calling ar.values[p][q], but needs to be used since it's defined as a pointer.