Added function to test surrounding locks.
Previously tested if at least one surrounding element was locked. New
function mintrue() tests for a defined minimum.

Setting the minimum to two still saves time compared to not locking
converged elements, but not as much as just one. Need to use difference
to compare error and see if it's worthwhile.